requirements.txt
# Web Framework
flask>=2.0.0
flask-cors>=4.0.0
gunicorn>=20.1.0
Werkzeug>=2.0.0

# LLM Related
llama-cpp-python

# Configuration
python-dotenv>=0.19.0

# Development Tools
black==23.3.0         # 代码格式化
ruff==0.0.270        # 代码检查
pytest>=7.0.0        # 测试框架

# Optional - 如果需要数据处理
numpy>=1.21.0
pandas>=1.3.0